pyproject.toml
[tool.poetry]
name = "image2ascii"
version = "0.1.0"
description = "Convert image to ascii easily and export to .txt or to an image format"
authors = ["Luis Papiernik <lpapiernik24@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.2.0"
numpy = "^1.23.3"
scipy = "^1.9.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
